chore(ui): restore pages under app/, fix Tailwind v4 PostCSS config, add tailwind.config.ts, move libs to app/lib, update UI shell and dataset pages
