+++ app/globals.css
@@ -37,7 +37,7 @@
     @apply inline-flex items-center justify-center rounded-lg px-4 py-2 text-sm font-medium transition-colors;
   }
   .btn-primary {
-    @apply btn text-white bg-sky-600 hover:bg-sky-700;
+    @apply inline-flex items-center justify-center rounded-lg px-4 py-2 text-sm font-medium transition-colors text-white bg-sky-600 hover:bg-sky-700;
   }
   .input {
     @apply rounded-lg border px-3 py-2 text-sm shadow-sm focus:outline-none focus:ring-2 focus:ring-sky-500/40;
