+++ prisma/seed.ts
@@ -21,3 +21,7 @@ main()
   })
 
 
+
+
+
+
